Do some work towards actually adding some of our actual sponsors
On the plus side, replaced some of the dummy content with real content. On the other hand, some of these logos could do with a little bit of... tlc? In addition, this isn't are most recent sponsor list. This is what is curently up on the current site (team1294.org). Both should be updated (see also FRC-1294/1294-Website#104).