Sponsor logos (#1)
* Do some work towards actually adding some of our actual sponsors

On the plus side, replaced some of the dummy content with real content. On the other hand, some of these logos could do with a little bit of... tlc? In addition, this isn't are most recent sponsor list. This is what is curently up on the current site (team1294.org). Both should be updated (see also FRC-1294/1294-Website#104).

* Delete .DS_Store

* Delete app.json

Heroku Review Apps + Jekyll = Huge Failure

* Extract slideshow script and move all scripts to the end of <body>

* Remove carousel

To be replaced with a still image

* Change logo

The "stripe-logo" technically hasn't been cleared yet...

* Move floating submenu because of the smaller logo

This really shouldn't rely on the width of a logo...

* Remove empty CSS rule

* Move the floating submenu because of the new height of the logo

* Add new sponsor logos