Get everything working again
Everything works locally on my machine (even though the theme still looks horrible...). Pages load properly and are in their own _pages folder. Styles are now in SCSS and take advantage of at least a few features of SASS. Updated and cleaned up the Gemfile and _config.yml a little.